added syncronized webserver interaction example with mocked test example

+++ src/main/java/com/coders51/rabbitmq/endpoint/PraticaEndpoint.java
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.client.RestTemplate;
 
 import com.coders51.rabbitmq.infra.pratica.Pratica;
 import com.coders51.rabbitmq.infra.pratica.PraticaService;
@@ -19,6 +20,9 @@ public class PraticaEndpoint {
 
 	@Autowired
 	PraticaService praticaService;
+
+	@Autowired
+    private RestTemplate restTemplate;
 
 	@GetMapping("/pratiche")
 	public List<Pratica> get() {
@@ -32,6 +36,12 @@ public class PraticaEndpoint {
 
 	@DeleteMapping("/pratiche/{id}")
 	public void delete(@PathVariable UUID id) {
+		// authorization test
+		String resp = restTemplate.getForObject("http://localhost:8081/auth.json", String.class);
+		if(resp.contains("false")) {
+			return;
+		}
+
 		praticaService.delete(id);
 	}
